feat: harden CI and release pipeline

+++ internal/adminapi/socket.go
@@ -61,5 +61,5 @@ func Listen(path string) (net.Listener, error) {
 
 func ownedByCurrentUser(info os.FileInfo) bool {
 	stat, ok := info.Sys().(*syscall.Stat_t)
-	return ok && stat.Uid == uint32(os.Getuid())
+	return ok && int64(stat.Uid) == int64(os.Getuid())
 }
